Routes constant and reuse modal

+++ frontend/src/pages/TicketManagerPage.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useState, useEffect } from "react";
 import { useNavigate } from 'react-router-dom';
 import { useTicketStore } from "../store/useTicketStore";
-import { Search, Plus, Trash, CheckCheck, Clock, Loader } from "lucide-react";
+import { Search, Plus, Trash, CheckCheck, Clock, Loader, MessageSquare, UserCheck2 } from "lucide-react";
 import { toastWarning } from "../components/ToastUtils";
 import TicketModal from "../components/TicketModal";
 import { useChatStore } from "../store/useChatStore";
@@ -11,18 +11,19 @@ import { DateTimeFormatter } from "../components/BasicUIComponents";
 import UnreadBadge from '../components/UnreadBadge';
 import ConfirmationModal from "../components/ConfirmationModal";
 import Select from "react-select";
+import { ROUTES } from "../constants/paths";
 
 const TicketManagerPage = () => {
   const { authUser, staffList, fetchStaffList } = useAuthStore();
   const { tickets, fetchAllTickets, setSelectedTicket, updateTicketStatus, filter, setFilter, isLoadingTickets,
-    statusList, filteredTickets, createTicket, updateTicketStaff, getCategories, deleteAllTickets, levelList, getTicketCountByStatus } = useTicketStore();
+    statusList, filteredTickets, createTicket, updateTicketStaff, getCategories, deleteAllTickets, levelList, getTicketCountByStatus, isTicketCreator } = useTicketStore();
 
   const { unreadCount, getUnreadCounts, latestMessages, getLatestMessages, formatLatestMessages } = useChatStore();
 
   const [searchTerm, setSearchTerm] = useState("");
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isDeleting, setIsDeleting] = useState(false);
   const [isModalStaffOpen, setIsModalStaffOpen] = useState(false);
+  const [isModalDeleteOpen, setIsModalDeleteOpen] = useState(false);
 
   useEffect(() => {
     fetchAllTickets();
@@ -37,8 +38,8 @@ const TicketManagerPage = () => {
   };
 
   const defaultAssignStaffForm = {
-    ticketId: null, 
-    ticketCreatorId: null, 
+    ticketId: null,
+    ticketCreatorId: null,
     staffId: null
   };
 
@@ -79,15 +80,11 @@ const TicketManagerPage = () => {
   };
 
   const handleDelete = async () => {
-    if (!window.confirm("Delete all tickets? \nThis is a permanent action.")) return;
-    setIsDeleting(true);
-
     try {
       await deleteAllTickets();
     } catch (error) {
       console.error(error);
     } finally {
-      setIsDeleting(false);
       fetchAllTickets();
     }
   };
@@ -129,14 +126,21 @@ const TicketManagerPage = () => {
     });
   }, [filteredStaffList, authUser._id]);
 
-  return (
-    <div className="p-6 pt-20 w-full mx-auto max-w-6xl space-y-4">
-      <div className="flex flex-row justify-between">
-        <h2 className="text-xl font-semibold">Chat Manager</h2>
+  const canManageTicket = (ticket) => {
+    return !isTicketCreator(ticket, authUser);
+  }
 
-        <div className="flex items-end justify-right space-x-3">
+  return (
+    <div className="p-6 pt-20 w-full mx-auto max-w-5xl space-y-4">
+      <div className="flex flex-row justify-between">
+        <div>
+          <h2 className="text-xl lg:text-2xl font-semibold">Chat Manager</h2>
+          <p className="text-sm text-gray-500">Manage chats with users</p>
+        </div>
+        
+        <div className="flex items-top justify-right space-x-3">
           <button
-            onClick={(e) => navigate('/ticket-chats')}
+            onClick={(e) => navigate(ROUTES.LIVE_CHAT)}
             className="btn flex p-1 rounded hover:bg-base-200 transition bg-emerald-400 cursor-pointer"
             title="Navigate to Live Chat"
           >
@@ -144,7 +148,7 @@ const TicketManagerPage = () => {
           </button>
 
           <button
-            onClick={(e) => handleDelete()}
+            onClick={(e) => setIsModalDeleteOpen(true)}
             className="btn flex p-1 rounded hover:bg-base-200 transition bg-red-400 cursor-pointer"
             title="Delete all tickets"
           >
@@ -152,19 +156,18 @@ const TicketManagerPage = () => {
             Delete All
           </button>
 
-          <button
+          {/* <button
             onClick={openCreateModal}
             className="btn flex gap-1 items-center btn-custom-primary"
             title="Create new ticket"
           >
-            {/* <Plus className="size-4" /> */}
             <span>New</span>
-          </button>
+          </button> */}
         </div>
       </div>
 
       {/* Tabs */}
-      <div className="flex gap-3 border-b pb-2 text-sm">
+      <div className="flex gap-3 bg-white rounded-lg shadow-sm p-2">
         {statusList.map((tab) => {
           const count = getTicketCountByStatus(tab);
 
@@ -172,20 +175,20 @@ const TicketManagerPage = () => {
             <button
               key={tab}
               onClick={() => setFilter(tab)}
-              className={`px-3 py-1 border-b-2 cursor-pointer transition-colors duration-200  ${filter === tab
+              className={`px-2 sm:px-3 py-1 border-b-2 cursor-pointer transition-colors duration-200 text-sm ${filter === tab
                 ? "border-blue-700 text-blue-700"
                 : "border-transparent text-gray-500 hover:text-blue-400"
                 }`}
             >
               {tab}
-              
+
               {/* Count badge */}
               <span
-                className={`text-xs ml-1 px-2 py-0.5 rounded-full font-medium
+                className={`text-xs ml-1 px-1.5 py-0.5 rounded-full font-medium
                 ${filter === tab
-                  ? "bg-blue-100 text-blue-700"
-                  : "bg-gray-200 text-gray-600"
-                }`}
+                    ? "bg-blue-100 text-blue-700"
+                    : "bg-gray-400 text-gray-100"
+                  }`}
               >
                 {count}
               </span>
@@ -204,24 +207,24 @@ const TicketManagerPage = () => {
         ) : (
           <>
             {visibleTickets.length > 0 ? (
-              <table className="table w-full">
+              <table className="table w-full table-bordered border-collapse">
                 <thead className="bg-gray-200 text-gray-700 uppercase text-sm">
                   <tr>
-                    <th className="px-4 py-3 text-left">Category</th>
-                    <th className="px-4 py-3 text-left">Level</th>
-                    <th className="px-4 py-3 text-left">Latest Message</th>
-                    <th className="px-4 py-3 text-left">Created By</th>
-                    <th className="px-4 py-3 text-left">Assigned To</th>
-                    <th className="px-4 py-3 text-left">Status</th>
-                    <th className="px-4 py-3 text-center">Actions</th>
+                    <th>Category</th>
+                    <th>Level</th>
+                    <th>Latest Message</th>
+                    <th>Created By</th>
+                    <th>Assigned To</th>
+                    <th className="min-w-[100px]">Status</th>
+                    <th>Actions</th>
                   </tr>
                 </thead>
                 <tbody>
                   {visibleTickets.map((ticket) => (
                     <tr key={ticket._id} className="hover:bg-gray-50">
                       <td>{ticket.category}</td>
-                      <td className="max-w-[150px] truncate whitespace-nowrap overflow-hidden" 
-                          title="">
+                      <td className="max-w-[150px] truncate whitespace-nowrap overflow-hidden"
+                        title="">
                         {ticket?.level}
                       </td>
                       <td className="max-w-[250px] truncate whitespace-nowrap overflow-hidden"
@@ -243,7 +246,7 @@ const TicketManagerPage = () => {
                             onClick={() => updateTicketStatus(ticket._id, "In Progress")}
                             title="Mark as In Progress"
                           >
-                            <Clock className="size-5" />
+                            <Clock className="size-4 lg:size-5" />
                           </button>
                         )}
                         {ticket.status !== "Resolved" && (
@@ -252,31 +255,38 @@ const TicketManagerPage = () => {
                             onClick={() => updateTicketStatus(ticket._id, "Resolved")}
                             title="Mark as Resolved"
                           >
-                            <CheckCheck className="size-5" />
+                            <CheckCheck className="size-4 lg:size-5" />
                           </button>
                         )}
                       </td>
                       <td className="text-center space-x-2 space-y-1">
-                        <div className="relative inline-block">
-                          <button className="btn btn-md btn-custom-primary-light"
-                            onClick={() => { handleChat(ticket) }}
-                            title="Chat"
-                          >
-                            Chat
-                          </button>
+                        {canManageTicket(ticket) &&
+                          <div>
+                            <div className="relative inline-block mr-2">
+                              <button className="btn btn-sm lg:btn-md btn-custom-primary-light mt-1"
+                                onClick={() => { handleChat(ticket) }}
+                                title="Open chat"
+                              >
+                                <MessageSquare className="size-5"/>
+                              </button>
 
-                          <UnreadBadge count={unreadCount[ticket._id]} className="absolute -top-2 -right-1" />
-                        </div>
+                              <UnreadBadge count={unreadCount[ticket._id]} className="absolute -top-2 -right-1" />
+                            </div>
 
-                        <button className="btn btn-md btn-custom-primary-light"
-                          onClick={() => {
-                            setAssignFormState({ ticketId: ticket._id, ticketCreatorId: ticket.userId._id, staffId: null });
-                            setIsModalStaffOpen(true);
-                          }}    // To-do (Put out a modal to set staff name)
-                          title="Assign Staff"
-                        >
-                          Assign
-                        </button>
+                            <button className="btn btn-sm lg:btn-md btn-custom-primary-light mt-1"
+                              onClick={() => {
+                                setAssignFormState({ ticketId: ticket._id, ticketCreatorId: ticket.userId._id, staffId: null });
+                                setIsModalStaffOpen(true);
+                              }}    // To-do (Put out a modal to set staff name)
+                              title="Assign Staff"
+                            >
+                              <UserCheck2 className="size-5"/>
+                            </button>
+                          </div>
+                        }
+                        {!canManageTicket(ticket) &&
+                          <span>Your chat</span>
+                        }
                       </td>
                     </tr>
                   ))}
@@ -284,7 +294,7 @@ const TicketManagerPage = () => {
               </table>
             ) : (
               <div className="text-center text-gray-500 py-6">
-                No tickets found.
+                <span>No chats yet.</span>
               </div>
             )}
           </>
@@ -324,6 +334,16 @@ const TicketManagerPage = () => {
           isSearchable={true}
         />
       </ConfirmationModal>
+
+      <ConfirmationModal
+        isOpen={isModalDeleteOpen}
+        onClose={() => setIsModalDeleteOpen(false)}
+        title="Confirmation"
+        children="Are you sure to delete all chats? This is a permanent action!"
+        primaryButton={{ label: "Delete All", onClick: handleDelete }}
+        primaryButtonStyle={"bg-red-500 border-red-500 hover:bg-red-600 text-white"}
+        secondaryButton={{ label: "Cancel", onClick: () => setIsModalDeleteOpen(false) }}
+      ></ConfirmationModal>
 
     </div>
   );
